Components renamed to index
