+++ src/components/About/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import Card from "react-md/lib/Cards/Card";
 import CardText from "react-md/lib/Cards/CardText";
-import UserLinks from "../UserLinks/UserLinks";
+import UserLinks from "../UserLinks";
 import config from "../../../data/SiteConfig";
 import "./About.scss";
