+++ src/components/Footer/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import Button from "react-md/lib/Buttons";
 import { Link } from "gatsby";
-import UserLinks from "../UserLinks/UserLinks";
+import UserLinks from "../UserLinks";
 import config from "../../../data/SiteConfig";
 import "./Footer.scss";
