+++ src/components/Navigation/index.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import NavigationDrawer from "react-md/lib/NavigationDrawers";
-import ToolbarActions from "../ToolbarActions/ToolbarActions";
-import Footer from "../Footer/Footer";
+import ToolbarActions from "../ToolbarActions";
+import Footer from "../Footer";
 import GetNavList from "./NavList";
 import "./Navigation.scss";
