+++ src/components/PostPreview/index.jsx
@@ -8,8 +8,8 @@ import FontIcon from "react-md/lib/FontIcons";
 import { Link } from "gatsby";
 import moment from "moment";
 import Media, { MediaOverlay } from "react-md/lib/Media";
-import PostTags from "../PostTags/PostTags";
-import PostCover from "../PostCover/PostCover";
+import PostTags from "../PostTags";
+import PostCover from "../PostCover";
 import config from "../../../data/SiteConfig";
 import "./PostPreview.scss";
