+++ src/components/ToolbarActions/index.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import UserLinks from "../UserLinks/UserLinks";
+import UserLinks from "../UserLinks";
 import "./ToolbarActions.scss";
 
 class Toolbar extends Component {
