+++ src/components/UserInfo/index.jsx
@@ -6,7 +6,7 @@ import Avatar from "react-md/lib/Avatars";
 import FontIcon from "react-md/lib/FontIcons";
 import IconSeparator from "react-md/lib/Helpers/IconSeparator";
 import { Follow } from "react-twitter-widgets";
-import UserLinks from "../UserLinks/UserLinks";
+import UserLinks from "../UserLinks";
 import "./UserInfo.scss";
 
 class UserInfo extends Component {
