+++ src/layout/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Helmet from "react-helmet";
 import "font-awesome/scss/font-awesome.scss";
-import Navigation from "../components/Navigation/Navigation";
+import Navigation from "../components/Navigation";
 import config from "../../data/SiteConfig";
 import "./index.scss";
 import "./global.scss";
