+++ src/pages/about.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import Helmet from "react-helmet";
 import Layout from "../layout";
-import About from "../components/About/About";
+import About from "../components/About";
 import config from "../../data/SiteConfig";
 
 class AboutPage extends Component {
