+++ src/templates/post.jsx
@@ -4,14 +4,14 @@ import { graphql } from "gatsby";
 import Card from "react-md/lib/Cards";
 import CardText from "react-md/lib/Cards/CardText";
 import Layout from "../layout";
-import UserInfo from "../components/UserInfo/UserInfo";
-import Disqus from "../components/Disqus/Disqus";
-import PostTags from "../components/PostTags/PostTags";
-import PostCover from "../components/PostCover/PostCover";
-import PostInfo from "../components/PostInfo/PostInfo";
-import SocialLinks from "../components/SocialLinks/SocialLinks";
-import PostSuggestions from "../components/PostSuggestions/PostSuggestions";
-import SEO from "../components/SEO/SEO";
+import UserInfo from "../components/UserInfo";
+import Disqus from "../components/Disqus";
+import PostTags from "../components/PostTags";
+import PostCover from "../components/PostCover";
+import PostInfo from "../components/PostInfo";
+import SocialLinks from "../components/SocialLinks";
+import PostSuggestions from "../components/PostSuggestions";
+import SEO from "../components/SEO";
 import config from "../../data/SiteConfig";
 import "./b16-tomorrow-dark.css";
 import "./post.scss";
